Changed the form structure so the selection are perserved despite re-rendering
